% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{InputData}
\alias{InputData}
\title{InputData}
\usage{
InputData(IniText = NULL, SetValue)
}
\arguments{
\item{IniText}{sets the initial message in the Input Entry Widget}

\item{SetValue}{function to save the Input Data in the desired variable
In the example the input data (text or number) is stored in the variable Value
Observe that the name of the variable is free while <<- Item CANNOT be changed}
}
\description{
InputData a dialog toplevel + entry widget to input values
 The widget returns text or numbers depending on the alphanumeric input given
}
\examples{
\dontrun{
    Value <- NULL
    InputData(function(Value <<- Item))
    cat("\n ==>", Value)
}
}
