% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSUtilitiesTK.r
\name{treeview_insert}
\alias{treeview_insert}
\title{treeview_insert}
\usage{
treeview_insert(widget, where = "end", items)
}
\arguments{
\item{widget}{widget where to add data}

\item{where}{is an integer or "end" indicating where to insert the item}

\item{items}{a data.frame containing the data to insert. items MUST have the same n.col of the widget}
}
\description{
adds the data.frame items to a treeview
}
