% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSBarPlotGUI.r
\name{XPSBarPlot}
\alias{XPSBarPlot}
\title{Function to generate BAR plots}
\usage{
XPSBarPlot()
}
\description{
XPSBarPlot allows a full control of the various
  parameters forplotting data. Through a user friendly interface it
  is possible to set colors, lines or symbols, their weight,  modify
  title, X,Y labels and their dimensions, add/modify legend
  annotate the plot.This GUI is based on the Lattice package.
}
\examples{
\dontrun{
	XPSBarPlot()
}
}
