% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPSRefineBaseLineTKGUI.r
\name{XPSRefineBaseLine}
\alias{XPSRefineBaseLine}
\title{XPSRefineBaseLine}
\usage{
XPSRefineBaseLine()
}
\description{
XPSMoveBaseLine modifies the BaseLine level and 
 limits of a given Coreline. New edge values are provided by
}
\examples{
\dontrun{
	XPSRefineBaseLine()
}
}
